+++ package.json
@@ -6,9 +6,12 @@
     "@testing-library/jest-dom": "^5.17.0",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
+    "dom": "^0.0.3",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
+    "react-router-dom": "^6.27.0",
     "react-scripts": "5.0.1",
+    "router": "^1.3.8",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
@@ -17,7 +20,6 @@
     "test": "react-scripts test",
     "eject": "react-scripts eject",
     "cypress": "cypress open"
-
   },
   "eslintConfig": {
     "extends": [
